+++ config/versiones.php
@@ -1,10 +1,10 @@
 <?php
 
 return [
-    'IEE'           => '1.21.0',
-    'BhagamApps'    => '1.20.0',
-    'User'          => '2.5.0',
+    'IEE'           => '1.21.1',
+    'BhagamApps'    => '1.20.1',
+    'User'          => '2.5.1',
     'Inventario'    => '2.15.1',
-    'Apps'          => '1.5.1',
+    'Apps'          => '1.5.2',
     'CrudGenerator' => '1.1.0',
 ];